Restructure frontend/backend classes to sub namespaces.
