+++ src/Frontend/RateItHybrid.php
@@ -14,7 +14,7 @@
  * @filesource
  */
 
-namespace Hofff\Contao\RateIt;
+namespace Hofff\Contao\RateIt\Frontend;
 
 use Contao\BackendTemplate;
 use Contao\FrontendTemplate;
